% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{positions.RTS0}
\alias{positions.RTS0}
\title{Positions of nwe time series}
\format{A data frame with two columns named \code{lon} (longitud), and \code{lat} (latitude)}
\usage{
positions.RTS0
}
\description{
Positions of nwe time series
}
\keyword{datasets}
