% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{positions.TS0}
\alias{positions.TS0}
\title{Positions of original time series}
\format{A data frame with two columns named \code{lon} (longitud), and \code{lat} (latitude), asociated to the time series}
\usage{
positions.TS0
}
\description{
Positions of original time series
}
\keyword{datasets}
