% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rts_clean.R
\name{rts_clean}
\alias{rts_clean}
\title{Cleanning Function for Time Series}
\usage{
rts_clean(TS, seasonality = 12)
}
\arguments{
\item{TS}{is a data frame of time series}

\item{seasonality}{either the character string "periodic" or the span (in lags) of the loess window for seasonal extraction}
}
\description{
This function clean a time series with missing values.
}
