% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rts_data_example.R
\name{rts_data_example}
\alias{rts_data_example}
\title{Data Example for GeoRTS (Geo-reconstruction of Time Series)}
\usage{
rts_data_example(m.ts = 8, nx.rts = 5, ny.rts = 5)
}
\value{
returns a list containing the original time series "TS0" (randomly generated), the positions and weights associated  with the original series (""positions.TS0" and "weights.TS0") and finally return the positions of the reconstruction of time series "positions.RTS0".
}
\description{
This function generates a group of random data to use geoRts function.
}
