% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rts_distrib.R
\name{rts_distrib}
\alias{rts_distrib}
\title{Empirical Distribution Function}
\usage{
rts_distrib(Xrand, x = Xrand)
}
\arguments{
\item{Xrand}{is a vector with observations of a random variable}

\item{x}{is a vector which will be evaluate the Empirical Distribution Function based on \code{Xrand}}
}
\value{
a vector which contain the values of Empirical Distribution Function evaluated in \code{x}
}
\description{
This function returns the Empirical Distribution Function based on a vector containig observations of a random variable
}
