% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rts_plotClean.R
\name{rts_plotClean}
\alias{rts_plotClean}
\title{Plot of Cleaned and original Time Series}
\usage{
rts_plotClean(TS, k = 1)
}
\arguments{
\item{TS}{is a object type ts wiht \code{ncol(.)>2} and It represents the original series that user establishes}

\item{k}{is the number of column (time serie) from TS to be plot.}
}
\value{
a plot with the original series and the reconstructed series are found
}
\description{
This function generates a plot with the original and reconstructed series
}
