% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rts_plotGroup.R
\name{rts_plotGroup}
\alias{rts_plotGroup}
\title{--------------- Plot of Time Series ----------------}
\usage{
rts_plotGroup(TS, RTS)
}
\arguments{
\item{TS}{is a object type ts wiht ncol mayor o equal to 2 and It represents the original series that user establishes}

\item{RTS}{is a object type ts that that is generated from the reconstruction of series}
}
\value{
a plot with the original series and the reconstructed series are found
}
\description{
This function generates a plot with the original and reconstructed series
}
