% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rts_simu.R
\name{rts_simu}
\alias{rts_simu}
\title{Simulation based on observations of a random variable}
\usage{
rts_simu(Xrand, n = 1)
}
\arguments{
\item{Xrand}{is a vector with observations of a random variable}

\item{n}{is the number of observations to be simulated}
}
\value{
a simulation of empirical distribucion of Xrand
}
\description{
This function simulates n observations of the random variable, based on the empirical distribution function of it's observations \code{Xrand}.
}
