% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{weights.TS0}
\alias{weights.TS0}
\title{Weights of original time series}
\format{A vector of length equal to number of rows of \code{positions.TS0}, asociate with the influence of original time series}
\usage{
weights.TS0
}
\description{
Weights of original time series
}
\keyword{datasets}
