% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoRts.R
\name{geoRts}
\alias{geoRts}
\title{Geo-reconstruction of Time Series}
\usage{
geoRts(TS, positions.TS, weights.TS = NULL, positions.RTS,
  weights.RTS = NULL, D = NULL)
}
\arguments{
\item{TS}{is a object of class "ts" (time series), with \code{ncol()>2}}

\item{positions.TS}{is a data frame that containing longitude and latitude asociated to time series in TS}

\item{weights.TS}{is a data frame that containing weights (of influence) of time series in TS.}

\item{positions.RTS}{is a data frame that containing values of longitude and latitude about each serie for build.}

\item{weights.RTS}{is a data frame that containing weights (of influence) of time series in RTS.}

\item{D}{is a distance matrix of all ditances. This should be entered by the user}
}
\value{
returns an object of class "ts", containing the reconstructed time series associate to RTS.positions
}
\description{
This function aproximate by interpolation the values of a time series (unknonw), based in the values of near (geographically) time series.
}
