% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/georts.R
\name{geoRts}
\alias{geoRts}
\title{This function build news tieme series through seires series established by the user}
\usage{
geoRts(TS, positions.TS, weights.TS = NULL, positions.RTS,
  weights.RTS = NULL, D = NULL)
}
\arguments{
\item{TS}{is a data frame of time series}

\item{positions.TS}{is a data frame that contain longitude and latitude}

\item{weights.TS}{is a data frame that contain values of each original serie.}

\item{positions.RTS}{is a data frame that contain values of longitude and latitude about each serie for build.}

\item{weights.RTS}{is a data frame that contain values of each series for build. This can be NULL}

\item{D}{is a distance matrix of all ditances. This should be entered by the user}
}
\value{
RTS new series
}
\description{
This function build news tieme series through seires series established by the user
}
