% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rts_data_example.R
\name{rts_data_example}
\alias{rts_data_example}
\title{Data Example for GeoRTS (Geo-reconstruction of Time Series)}
\usage{
rts_data_example(m.ts = 8, nx.rts = 5, ny.rts = 5)
}
\arguments{
\item{m.ts}{is a interger that represents the number of original time series "TS",  \code{m.ts=dim(TS)[2]}}

\item{nx.rts}{is a interger that represents the number of points of longitude \code{log}(new positions) of the grid.}

\item{ny.rts}{is a interger that represents the number of points of latitude \code{lat}(new positions) of the grid.}
}
\value{
returns a list containing the original time series "TS0" (randomly generated), the positions and weights associated  with the original series (""positions.TS0" and "weights.TS0") and finally return tha positions of the reconstruction of time series "positions.RTS0".
}
\description{
This function generates a group of random data to use geoRts function.
}
