% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rts_map.R
\name{rts_map}
\alias{rts_map}
\title{Map representation of Geographical Time Series}
\usage{
rts_map(positions.TS, positions.RTS, scale = 1, weights.TS = NULL,
  weights.RTS = NULL, names.TS = NULL, names.RTS = NULL)
}
\arguments{
\item{positions.TS}{is a data frame that contain longitude and latitude of time series}

\item{positions.RTS}{is a data frame that contain longitude and latitude of reconstructed time series}

\item{weights.TS}{is a vector that contain the weights of positions}

\item{weights.RTS}{is a vector that contain the weights of positions}

\item{names.TS}{is a vector that contain of names of positions}

\item{names.RTS}{is a vector that contain of names of positions}

\item{TS}{is a ts object}

\item{RTS}{is a ts object}
}
\value{
returns a object of class "leaflet", that contain a animated map with points that represents the time series in it's geographical positions
}
\description{
This function build an dynamic map, which represents the geographical time series
}
