% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rts_map_raster.R
\name{rts_map_raster}
\alias{rts_map_raster}
\title{Map representation of Geographical Time Series (on a grid)}
\usage{
rts_map_raster(TS, positions.TS, weights.TS = NULL, type = c("2D",
  "3D", "2D+3D", "2D-dynamic", "3D-dynamic"), k = 1, fpss = 5,
  windowsize = c(700, 700), save.plot = TRUE, file.name = "plot",
  nx.grid = 10 * dim(positions.TS)[1], ny.grid = 10 *
  dim(positions.TS)[1], ...)
}
\arguments{
\item{TS}{is a object of class "ts" (time series), with \code{ncol()>2}}

\item{positions.TS}{is a data frame that containing longitude and latitude asociated to time series in TS.}

\item{weights.TS}{is a data frame that containing weights (of influence) of time series in TS.}

\item{type}{is the type of plot to be shown, available options are: \code{"2D","3D","2D+3D", "2D-dynamic","3D-dynamic"}}

\item{k}{Represent the number of time serie (column of \code{TS}). Is used only if \code{type} parameter is set as one of following options: \code{"2D","3D"} or \code{"2D+3D"}}

\item{fpss}{is the number of frames per second. Is only used for types: \code{"2D-dynamic"} or \code{"3D-dynamic"}}

\item{windowsize}{is a vector that contain width and height}

\item{save.plot}{set as TRUE for save the plot}

\item{file.name}{is a character for name of saved plot}

\item{...}{Arguments passed to \code{geoRts} or \code{plot_gg} function, from  \code{rayshader} package}
}
\value{
returns a object of class "leaflet", that contain a animated map with points that represents the time series in it's geographical positions
}
\description{
This function build an dynamic map, which represents the geographical time series on a grid
}
